+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to
   configure the HD path of the faucet accounts ([#832]).
 - @cosmjs/cosmwasm-stargate: Add field `ibcPortId` to `Contract` ([#836]).
 - @cosmjs/stargate: Add `GovExtension` for query client.
+- @cosmjs/stargate: Add support for `MsgSubmitProposal` and `MsgVote`.
 
 [#832]: https://github.com/cosmos/cosmjs/issues/832
 [#836]: https://github.com/cosmos/cosmjs/issues/836